+++ src/Oro/Bundle/CommerceMenuBundle/Menu/ContextProvider/CustomerMenuContextProvider.php
@@ -7,6 +7,9 @@ use Oro\Bundle\CustomerBundle\Provider\ScopeCustomerCriteriaProvider;
 use Oro\Bundle\WebsiteBundle\Manager\WebsiteManager;
 use Oro\Bundle\WebsiteBundle\Provider\ScopeCriteriaProvider;
 
+/**
+ * Adds the current customer and the default website to the customer menu context.
+ */
 class CustomerMenuContextProvider implements CustomerMenuContextProviderInterface
 {
     /** @var WebsiteManager */
@@ -27,7 +30,7 @@ class CustomerMenuContextProvider implements CustomerMenuContextProviderInterfac
     {
         return [
             [
-                ScopeCustomerCriteriaProvider::ACCOUNT => $customer->getId(),
+                ScopeCustomerCriteriaProvider::CUSTOMER => $customer->getId(),
                 ScopeCriteriaProvider::WEBSITE => $this->websiteManager->getDefaultWebsite()->getId()
             ]
         ];